pyproject.toml
[tool.poetry]
name = "purplestore-backend-test"
version = "0.1.0"
description = ""
authors = ["이한영 <lhy@purple-nest.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "<4"
djangorestframework = "^3.13.1"
django-extensions = "^3.1.5"
model-bakery = "^1.4.0"

[tool.poetry.dev-dependencies]

[tool.black]
line-length = 130

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
